don't crash!
[git-p4: depot-paths = "//depot/qt/main/": change = 167481]

+++ src/gui/painting/qpaintengine_x11_p.h
@@ -43,7 +43,7 @@ struct qt_XPoint {
 
 struct qt_float_point
 {
-    float x, y;
+    qreal x, y;
     operator qt_XPoint() const
     {
         qt_XPoint pt = { static_cast<short>(qRound(x)),
